error that has to do with host with host_id 0

+++ Lab07/packet.c
@@ -6,7 +6,6 @@
 
 #include "packet.h"
 #include "net.h"
-#include "host.h"
 
 
 void packet_send(struct net_port *port, struct packet *p) {
@@ -22,11 +21,6 @@ void packet_send(struct net_port *port, struct packet *p) {
             msg[i + 4] = p->payload[i];
         }
         write(port->pipe_send_fd, msg, p->length + 4);
-//printf("PACKET SEND, src=%d dst=%d p-src=%d p-dst=%d\n", 
-//		(int) msg[0], 
-//		(int) msg[1], 
-//		(int) p->src, 
-//		(int) p->dst);
     }
 
     return;
@@ -38,7 +32,7 @@ int packet_recv(struct net_port *port, struct packet *p) {
     int i;
 
     if (port->type == PIPE) {
-        n = read(port->pipe_recv_fd, msg, PAYLOAD_MAX + 4);
+        n = (int) read(port->pipe_recv_fd, msg, PAYLOAD_MAX + 4);
         if (n > 0) {
             p->src = (char) msg[0];
             p->dst = (char) msg[1];
@@ -48,11 +42,6 @@ int packet_recv(struct net_port *port, struct packet *p) {
                 p->payload[i] = msg[i + 4];
             }
 
-// printf("PACKET RECV, src=%d dst=%d p-src=%d p-dst=%d\n", 
-//		(int) msg[0], 
-//		(int) msg[1], 
-//		(int) p->src, 
-//		(int) p->dst);
         }
     }
